Implement user management features: create UserController, add user role update functionality, integrate AuthPermission middleware, and update routes and views for user management.

+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\AuthController;
 use App\Http\Controllers\NavetteController;
 use App\Http\Controllers\RoleController;
 use App\Http\Controllers\StatisticController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 // Main landing page route
@@ -41,26 +42,30 @@ Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
 
 
 // Home page for client users (role_id = 1)
-Route::get('/navettes/home',[NavetteController::class, 'index'] );
+Route::get('/navettes/home', [NavetteController::class, 'index'])->middleware('AuthPermission:view_navettes')->name('navettes.home');
 
 // Route for searching navettes
-Route::get('/navettes/search', [NavetteController::class, 'search'])->name('navettes.search');
+Route::get('/navettes/search', [NavetteController::class, 'search'])->middleware('AuthPermission:search_navettes')->name('navettes.search');
 
 // Route for displaying a specific navette
-Route::get('/navettes/{navette}', [NavetteController::class, 'show'])->name('navettes.show');
+Route::get('/navettes/{navette}', [NavetteController::class, 'show'])->middleware('AuthPermission:view_navette_details')->name('navettes.show');
 
 
 
 // Admin dashboard
-Route::get('/admin/dashboard', [StatisticController::class, 'index'])->name('dashboard');
+Route::get('/admin/dashboard', [StatisticController::class, 'index'])->middleware('AuthPermission:dashboard')->name('dashboard');
 
-Route::get('/admin/roles/index', [RoleController::class, 'index'])->name('roles.index');
+// Role management routes
+Route::get('/admin/roles/index', [RoleController::class, 'index'])->middleware('AuthPermission:roles.index')->name('roles.index');
 
-Route::get('/admin/roles/create', [RoleController::class, 'create'])->name('roles.create');
-Route::post('/admin/roles/store', [RoleController::class, 'store'])->name('roles.store');
+Route::get('/admin/roles/create', [RoleController::class, 'create'])->middleware('AuthPermission:roles.create')->name('roles.create');
+Route::post('/admin/roles/store', [RoleController::class, 'store'])->middleware('AuthPermission:roles.store')->name('roles.store');
 
-Route::get('/admin/roles/edit/{role}', [RoleController::class, 'edit'])->name('roles.edit');
-Route::put('/admin/roles/update/{role}', [RoleController::class, 'update'])->name('roles.update');
+Route::get('/admin/roles/edit/{role}', [RoleController::class, 'edit'])->middleware('AuthPermission:roles.edit')->name('roles.edit');
+Route::put('/admin/roles/update/{role}', [RoleController::class, 'update'])->middleware('AuthPermission:roles.update')->name('roles.update');
+
+Route::delete('/admin/roles/destroy{role}', [RoleController::class, 'destroy'])->middleware('AuthPermission:roles.destroy')->name('roles.destroy');
 
 
-Route::delete('/admin/roles/destroy{role}', [RoleController::class, 'destroy'])->name('roles.destroy');
+Route::get('/admin/users/index', [UserController::class, 'index'])->name('users.index');
+Route::patch('/admin/users/update-role/{role_id}', [UserController::class, 'update'])->name('admin.users.update-role');
